[bees] Chat UI in Surface Pane (Bento Grid Layout)
## What
Extracts chat UI from the Detail tab into a standalone `<bees-chat-panel>` component and composes it as a built-in item in the Surface pane using a CSS grid bento box layout. Also fixes a bug where choice responses were never logged to `chat_log.json`.

## Why
Chat and surface items are conceptually part of the same presentation — the agent's curated output. Unifying them in a single pane with a structured grid layout creates a more cohesive experience (chat on the left, surface content on the right) and sets up the infrastructure for richer bento-style layouts.

## Changes

### Hivetool UI
- **[NEW] `chat-panel.ts`** — extracted `<bees-chat-panel>` combining chat log + input UI (text reply, choice cards). Renders markdown in chat messages and agent prompts. Deduplicates the last agent message when the input panel is showing.
- **`surface-pane.ts`** — bento grid layout owner. Composes chat panel alongside bundles and surface-view. Chat takes left half when present; content fills right half or full width.
- **`ticket-detail.ts`** — removed ~400 lines of chat UI code (log, reply/choice forms, related state and styles). Non-interactive suspend fallback preserved.
- **`ticket-pane.ts`** — Surface tab now shown when surface OR chat content exists. Passes `mutationClient` to surface pane. Fixed `overflow: auto` on tab body for Detail scrolling.
- **Exported `hasChatContent(ticket)`** utility for tab visibility probing.

### Backend
- **`task_runner.py`** — `resume_task` now logs choice responses (`selectedIds`) to `chat_log.json`, not just text replies.

## Testing
- Select a task with chat history but no surface → Surface tab appears, chat renders full-width
- Select a task with surface + chat → two-column bento layout, chat left, items right
- Select a task with surface only → items flow full-width
- Reply to a suspended task (text and choice) → both logged to `chat_log.json`
- Detail tab scrolls correctly
