+++ packages/bees/hive/config/SYSTEM.yaml
@@ -7,4 +7,4 @@
 #               exists yet.
 title: Opal
 description: Personal assistant
-root: live-session
+root: opie
