+++ packages/bees/hivetool/src/ui/ticket-detail.ts
@@ -32,55 +32,6 @@ class BeesTicketDetail extends SignalWatcher(LitElement) {
     sharedStyles,
     jsonTreeStyles,
     css`
-      /* ── Chat log ── */
-      .chat-log {
-        display: flex;
-        flex-direction: column;
-        gap: 4px;
-        padding: 8px 12px;
-        max-height: 500px;
-        overflow-y: auto;
-      }
-
-      .chat-turn {
-        border-radius: 8px;
-        padding: 10px 14px;
-        font-size: 0.8rem;
-        line-height: 1.5;
-      }
-
-      .chat-turn.user {
-        background: #1e293b;
-        border: 1px solid #334155;
-      }
-
-      .chat-turn.agent {
-        background: #111827;
-        border: 1px solid #1e293b;
-      }
-
-      .chat-role {
-        font-size: 0.65rem;
-        font-weight: 700;
-        text-transform: uppercase;
-        letter-spacing: 0.06em;
-        margin-bottom: 4px;
-      }
-
-      .chat-turn.user .chat-role {
-        color: #60a5fa;
-      }
-
-      .chat-turn.agent .chat-role {
-        color: #a78bfa;
-      }
-
-      .chat-text {
-        color: #e2e8f0;
-        white-space: pre-wrap;
-        word-break: break-word;
-      }
-
       /* ── File tree ── */
       .file-tree {
         font-family: "Google Mono", "Roboto Mono", monospace;
@@ -164,144 +115,6 @@ class BeesTicketDetail extends SignalWatcher(LitElement) {
         color: #cbd5e1;
         white-space: pre-wrap;
         word-break: break-word;
-      }
-
-      /* ── Response form ── */
-
-      .response-form {
-        display: flex;
-        flex-direction: column;
-        gap: 12px;
-      }
-
-      .agent-prompt {
-        padding: 12px 16px;
-        background: #111827;
-        border: 1px solid #1e293b;
-        border-radius: 8px;
-        color: #e2e8f0;
-        font-size: 0.85rem;
-        line-height: 1.6;
-        white-space: pre-wrap;
-      }
-
-      .reply-textarea {
-        width: 100%;
-        min-height: 80px;
-        padding: 10px 12px;
-        background: #0b0c0f;
-        border: 1px solid #334155;
-        border-radius: 6px;
-        color: #e2e8f0;
-        font-family: inherit;
-        font-size: 0.85rem;
-        resize: vertical;
-        outline: none;
-        transition: border-color 0.15s;
-      }
-
-      .reply-textarea:focus {
-        border-color: #3b82f6;
-      }
-
-      .reply-textarea::placeholder {
-        color: #475569;
-      }
-
-      .reply-actions {
-        display: flex;
-        justify-content: flex-end;
-        gap: 8px;
-      }
-
-      .send-btn {
-        padding: 6px 16px;
-        font-size: 0.8rem;
-        font-weight: 600;
-        background: #1d4ed8;
-        color: #dbeafe;
-        border: 1px solid #2563eb;
-        border-radius: 6px;
-        cursor: pointer;
-        font-family: inherit;
-        transition: all 0.15s;
-      }
-
-      .send-btn:hover {
-        background: #2563eb;
-        color: #fff;
-      }
-
-      .send-btn:disabled {
-        opacity: 0.4;
-        cursor: not-allowed;
-      }
-
-      /* ── Choice cards ── */
-
-      .choices-grid {
-        display: flex;
-        flex-direction: column;
-        gap: 6px;
-      }
-
-      .choice-card {
-        display: flex;
-        align-items: center;
-        gap: 10px;
-        padding: 10px 14px;
-        background: #111827;
-        border: 1px solid #1e293b;
-        border-radius: 8px;
-        cursor: pointer;
-        transition: all 0.15s;
-        font-size: 0.8rem;
-        color: #e2e8f0;
-      }
-
-      .choice-card:hover {
-        background: #1e293b;
-        border-color: #334155;
-      }
-
-      .choice-card.selected {
-        background: #1e3a5f;
-        border-color: #3b82f6;
-      }
-
-      .choice-indicator {
-        width: 16px;
-        height: 16px;
-        border-radius: 50%;
-        border: 2px solid #475569;
-        flex-shrink: 0;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        transition: all 0.15s;
-      }
-
-      .choice-card.selected .choice-indicator {
-        border-color: #3b82f6;
-        background: #3b82f6;
-      }
-
-      .choice-indicator::after {
-        content: "";
-        width: 6px;
-        height: 6px;
-        border-radius: 50%;
-        background: transparent;
-        transition: background 0.15s;
-      }
-
-      .choice-card.selected .choice-indicator::after {
-        background: #fff;
-      }
-
-      .choice-label {
-        flex: 1;
-        line-height: 1.4;
       }
 
       /* ── Live session panel ── */
@@ -427,11 +240,6 @@ class BeesTicketDetail extends SignalWatcher(LitElement) {
   @state() accessor fileTree: FileTreeNode[] = [];
   @state() accessor fileContents: Record<string, string | null> = {};
 
-  // ── Response state ──
-  @state() accessor replyText = "";
-  @state() accessor selectedChoiceIds = new Set<string>();
-  @state() accessor responding = false;
-
   // Live session state is owned by TicketStore.activeConnection —
   // no local client reference needed.
 
@@ -453,9 +261,7 @@ class BeesTicketDetail extends SignalWatcher(LitElement) {
       this.#treeLoadedFor = ticket.id;
     }
 
-    const chatHistory = (ticket.chat_history ?? []).filter(
-      (m) => m.text.trim() !== ""
-    );
+
 
     return html`
       <div class="timeline">
@@ -488,29 +294,7 @@ class BeesTicketDetail extends SignalWatcher(LitElement) {
               </div>
             `
           : nothing}
-        ${chatHistory.length > 0
-          ? html`
-              <div class="block">
-                <div class="block-header">
-                  Chat (${chatHistory.length} messages)
-                </div>
-                <div class="chat-log">
-                  ${chatHistory.map(
-                    (m) => html`
-                      <div
-                        class="chat-turn ${m.role === "user"
-                          ? "user"
-                          : "agent"}"
-                      >
-                        <div class="chat-role">${m.role}</div>
-                        <div class="chat-text">${m.text}</div>
-                      </div>
-                    `
-                  )}
-                </div>
-              </div>
-            `
-          : nothing}
+
         ${ticket.outcome
           ? html`
               <div class="block outcome">
@@ -538,7 +322,7 @@ class BeesTicketDetail extends SignalWatcher(LitElement) {
           ? this.renderLiveSessionPanel(ticket.id)
           : nothing}
         ${ticket.status === "suspended" && ticket.suspend_event
-          ? this.renderSuspendedBlock(ticket.id, ticket.suspend_event, ticket.assignee)
+          ? this.renderSuspendedBlock(ticket.suspend_event)
           : nothing}
         ${ticket.tags && ticket.tags.length > 0
           ? html`
@@ -702,41 +486,17 @@ class BeesTicketDetail extends SignalWatcher(LitElement) {
     };
   }
 
-  // ── Suspend / Response ──
+  // ── Suspend ──
 
   /**
-   * Render the suspended block — interactive response form for user-facing
-   * suspensions, JSON tree for everything else.
+   * Render the suspended block — raw JSON tree for non-interactive
+   * suspensions. User-facing interactive input (chat) is handled by
+   * `<bees-chat-panel>` in the surface pane.
    */
   private renderSuspendedBlock(
-    ticketId: string,
-    suspendEvent: Record<string, unknown>,
-    assignee?: string
+    suspendEvent: Record<string, unknown>
   ) {
     const functionName = suspendEvent.function_name as string | undefined;
-    const isUserFacing =
-      assignee === "user" &&
-      functionName !== "chat_await_context_update";
-
-    // For user-facing suspensions, show interactive response UI
-    // only when the box is actively listening for mutations.
-    if (isUserFacing && this.mutationClient?.boxActive.get()) {
-      const waitForInput = suspendEvent.waitForInput as
-        | Record<string, unknown>
-        | undefined;
-      const waitForChoice = suspendEvent.waitForChoice as
-        | Record<string, unknown>
-        | undefined;
-
-      if (waitForInput) {
-        return this.renderReplyForm(ticketId, waitForInput);
-      }
-      if (waitForChoice) {
-        return this.renderChoiceForm(ticketId, waitForChoice);
-      }
-    }
-
-    // Fallback: raw JSON tree for non-interactive or unknown events.
     const label =
       functionName === "chat_await_context_update"
         ? "Waiting for Event"
@@ -751,166 +511,6 @@ class BeesTicketDetail extends SignalWatcher(LitElement) {
         </div>
       </div>
     `;
-  }
-
-  /** Extract displayable text from an LLMContent prompt. */
-  private extractPromptText(prompt: unknown): string {
-    if (!prompt || typeof prompt !== "object") return "";
-    const p = prompt as { parts?: Array<{ text?: string }> };
-    if (!Array.isArray(p.parts)) return "";
-    return p.parts.map((part) => part.text ?? "").join("");
-  }
-
-  /** Interactive text reply form for waitForInput suspensions. */
-  private renderReplyForm(
-    ticketId: string,
-    waitForInput: Record<string, unknown>
-  ) {
-    const promptText = this.extractPromptText(waitForInput.prompt);
-
-    return html`
-      <div class="block">
-        <div class="block-header">Reply</div>
-        <div class="block-content">
-          <div class="response-form">
-            ${promptText
-              ? html`<div class="agent-prompt">${promptText}</div>`
-              : nothing}
-            <textarea
-              class="reply-textarea"
-              placeholder="Type your reply…"
-              .value=${this.replyText}
-              ?disabled=${this.responding}
-              @input=${(e: Event) => {
-                this.replyText = (e.target as HTMLTextAreaElement).value;
-              }}
-              @keydown=${(e: KeyboardEvent) => {
-                if (e.key === "Enter" && (e.metaKey || e.ctrlKey)) {
-                  e.preventDefault();
-                  if (this.replyText.trim()) {
-                    this.handleTextReply(ticketId);
-                  }
-                }
-              }}
-            ></textarea>
-            <div class="reply-actions">
-              <button
-                class="send-btn"
-                ?disabled=${!this.replyText.trim() || this.responding}
-                @click=${() => this.handleTextReply(ticketId)}
-              >
-                ${this.responding ? "Sending…" : "Send"}
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    `;
-  }
-
-  /** Interactive choice selection form for waitForChoice suspensions. */
-  private renderChoiceForm(
-    ticketId: string,
-    waitForChoice: Record<string, unknown>
-  ) {
-    const promptText = this.extractPromptText(waitForChoice.prompt);
-    const choices = (waitForChoice.choices ?? []) as Array<{
-      id: string;
-      content?: { parts?: Array<{ text?: string }> };
-    }>;
-    const selectionMode =
-      (waitForChoice.selectionMode as string) ?? "single";
-
-    return html`
-      <div class="block">
-        <div class="block-header">Choose</div>
-        <div class="block-content">
-          <div class="response-form">
-            ${promptText
-              ? html`<div class="agent-prompt">${promptText}</div>`
-              : nothing}
-            <div class="choices-grid">
-              ${choices.map((choice) => {
-                const selected = this.selectedChoiceIds.has(choice.id);
-                const label = this.extractPromptText(choice.content) || choice.id;
-                return html`
-                  <div
-                    class="choice-card ${selected ? "selected" : ""}"
-                    @click=${() =>
-                      this.toggleChoice(choice.id, selectionMode)}
-                  >
-                    <div class="choice-indicator"></div>
-                    <span class="choice-label">${label}</span>
-                  </div>
-                `;
-              })}
-            </div>
-            <div class="reply-actions">
-              <button
-                class="send-btn"
-                ?disabled=${this.selectedChoiceIds.size === 0 ||
-                  this.responding}
-                @click=${() => this.handleChoiceReply(ticketId)}
-              >
-                ${this.responding ? "Sending…" : "Confirm"}
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    `;
-  }
-
-  private toggleChoice(choiceId: string, mode: string) {
-    const next = new Set(this.selectedChoiceIds);
-    if (mode === "single") {
-      // Single select: toggle or replace.
-      if (next.has(choiceId)) {
-        next.clear();
-      } else {
-        next.clear();
-        next.add(choiceId);
-      }
-    } else {
-      // Multiple select: toggle.
-      if (next.has(choiceId)) {
-        next.delete(choiceId);
-      } else {
-        next.add(choiceId);
-      }
-    }
-    this.selectedChoiceIds = next;
-  }
-
-  private async handleTextReply(ticketId: string) {
-    const text = this.replyText.trim();
-    if (!text || !this.mutationClient) return;
-
-    this.responding = true;
-    try {
-      await this.mutationClient.respondToTask(ticketId, { text });
-      this.replyText = "";
-    } catch (e) {
-      console.error("Failed to send reply:", e);
-    } finally {
-      this.responding = false;
-    }
-  }
-
-  private async handleChoiceReply(ticketId: string) {
-    if (this.selectedChoiceIds.size === 0 || !this.mutationClient) return;
-
-    this.responding = true;
-    try {
-      await this.mutationClient.respondToTask(ticketId, {
-        selectedIds: [...this.selectedChoiceIds],
-      });
-      this.selectedChoiceIds = new Set();
-    } catch (e) {
-      console.error("Failed to send choice:", e);
-    } finally {
-      this.responding = false;
-    }
   }
 
   // ── Live session ──
